requirements.txt
Flask==0.10.1
Flask-Cors==2.1.0
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
itsdangerous==0.24
nose==1.3.7
numpy==1.9.2
oauthlib==1.0.3
pandas==0.16.2
peewee==2.6.3
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
scikit-learn==0.16.1
scipy==0.16.0
six==1.9.0
sklearn==0.0
tweepy==3.4.0
tweetpy==0.1
wheel==0.24.0
wsgiref==0.1.2
pyyaml
django
djangorestframework
markdown
django-filter
